jk refactored into ViewTeam component
